Add management command for batch processing items
Facility List Items can be batch processed for parsing,
geocoding, or matching. Given an optional environment
variable AWS_BATCH_JOB_ARRAY_INDEX, the batch may be
limited the matching row_index row. Otherwise it runs
on all the items for a given list.

The output includes the number of successes and failures
for the operation.