Add a Rubota data fetch script to the research directory
We want to use the Rubota data for model training purposes. The legacy
application includes a script to download this data, but it was failing. This
new version corrects the failure, removes lint, removes code that we do not
need, and limits the output columns to country, name, and address.